- [ZBX-65] fixed "On massupdate sideeffect all "application" entries are cleaned from all affected items." issue (Artem)
git-svn-id: svn://svn.zabbix.com/branches/1.4@4771 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ frontends/php/include/items.inc.php
@@ -423,13 +423,15 @@
 
 		DBexecute("update items set lastlogsize=0 where itemid=$itemid and key_<>".zbx_dbstr($key));
 
-		$result = DBexecute("delete from items_applications where itemid=$itemid");
-		foreach($applications as $appid)
-		{
-			$itemappid=get_dbid("items_applications","itemappid");
-			DBexecute("insert into items_applications (itemappid,itemid,applicationid) values($itemappid,".$itemid.",".$appid.")");
+		if(isset($_REQUEST['applications_visible'])){	
+			$result = DBexecute("delete from items_applications where itemid=$itemid");
+			foreach($applications as $appid)
+			{
+				$itemappid=get_dbid("items_applications","itemappid");
+				DBexecute("insert into items_applications (itemappid,itemid,applicationid) values($itemappid,".$itemid.",".$appid.")");
+			}
 		}
-
+		
 		$result=DBexecute(
 			"update items set description=".zbx_dbstr($description).",key_=".zbx_dbstr($key).",".
 			"hostid=$hostid,delay=$delay,history=$history,nextcheck=0,type=$type,status=".$status.','.
@@ -492,7 +494,8 @@
 					"trends"		=> array('template' => 1 , 'httptest' => 1),
 					"logtimefmt"		=> array(),
 					"valuemapid"		=> array('httptest' => 1),
-					"delay_flex"		=> array());
+					"delay_flex"		=> array()
+					);
 
 		$item_data = get_item_by_itemid($itemid);
